backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
openai==1.51.0
python-dotenv==1.0.1

